# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Slide Goose",
        "contact": {},
        "license": {
            "name": "GPL-3.0-or-later",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        },
        "version": "alpha"
    },
    "basePath": "/api",
    "paths": {
        "/audio/{audioId}/tag": {
            "post": {
                "description": "return status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tag"
                ],
                "summary": "set audio Tag",
                "parameters": [
                    {
                        "description": "tag",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/router.setTagForm"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "audio id",
                        "name": "audioId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/router.JSONResult"
                        }
                    }
                }
            }
        },
        "/clip/{clipId}/tag": {
            "post": {
                "description": "return status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tag"
                ],
                "summary": "set clip Tag",
                "parameters": [
                    {
                        "description": "tag",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/router.setTagForm"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "clip id",
                        "name": "clipId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/router.JSONResult"
                        }
                    }
                }
            }
        },
        "/file": {
            "post": {
                "description": "return number of added files",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "file"
                ],
                "summary": "add file or files",
                "parameters": [
                    {
                        "description": "files",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/router.filesForm"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/picture/{pictureId}/tag": {
            "post": {
                "description": "return status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tag"
                ],
                "summary": "set picture Tag",
                "parameters": [
                    {
                        "description": "tag",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/router.setTagForm"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "picture id",
                        "name": "pictureId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/router.JSONResult"
                        }
                    }
                }
            }
        },
        "/tag": {
            "post": {
                "description": "return status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tag"
                ],
                "summary": "add Tag",
                "parameters": [
                    {
                        "description": "tag",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/router.addTagForm"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/router.JSONResult"
                        }
                    }
                }
            }
        },
        "/video/{videoId}/tag": {
            "post": {
                "description": "return status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tag"
                ],
                "summary": "set video Tag",
                "parameters": [
                    {
                        "description": "tag",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/router.setTagForm"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "video id",
                        "name": "videoId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/router.JSONResult"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "router.JSONResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "router.addTagForm": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "router.filesForm": {
            "type": "object",
            "properties": {
                "files": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "router.setTagForm": {
            "type": "object",
            "properties": {
                "tag": {
                    "type": "integer"
                }
            }
        }
    }
}